apps/assistant-streamlit/requirements.txt
streamlit>=1.42.0
websockets>=14.0
httpx>=0.28.0
pandas>=2.0.0
python-dotenv>=1.0.0